+++ dandy/debug/__init__.py
@@ -1,1 +1,7 @@
 from dandy.debug.debug import DebugRecorder
+from dandy.debug.decorators import debug_recorder_to_html
+
+__all__ = [
+    'DebugRecorder',
+    'debug_recorder_to_html'
+]