+++ example/main.py
@@ -1,13 +1,15 @@
 import logging
+import os
 
 from book.intelligence.workflow import BookWorkflow
-from dandy.core.cache import SqliteCache
+from dandy.cache import SqliteCache
 from dandy.core.exceptions import DandyException
 from dandy.debug import DebugRecorder
 
 if __name__ == '__main__':
     try:
-        # SqliteCache.clear('example')
+        if os.getenv('DEBUG'):
+            SqliteCache.clear('example')
 
         logging.basicConfig(
             level=logging.INFO,