+++ docs/changelog/changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## v0.XX.X
+## v0.XX.X - Coming Soon
 
 ### Fixes